requirements.txt
streamlit>=1.30.0
beautifulsoup4>=4.12.0
requests>=2.31.0
chromadb>=0.4.22
sentence-transformers>=2.3.0
google-generativeai>=0.3.2
python-dotenv>=1.0.0
langchain-text-splitters>=0.0.1